server: move from js to ts
