+++ .gitignore
@@ -1,4 +1,7 @@
 node_modules/
+dist/
 .yarn/cache
 .yarn/unplugged
 .yarn/build-state.yml
+yarn-error.log
+
