+++ package.json
@@ -1,10 +1,19 @@
 {
   "name": "kaka",
   "version": "1.0.0",
-  "main": "index.js",
   "author": "Mayant Mukul <mayantmukul@hotmail.com>",
   "license": "MIT",
+  "scripts": {
+    "start": "npx tsc && node dist/index.js",
+    "build": "npx tsc",
+    "clean": "rm -rf dist"
+  },
   "dependencies": {
     "discord.js": "^12.1.1"
+  },
+  "devDependencies": {
+    "@types/node": "^15.3.0",
+    "eslint": "^7.26.0",
+    "typescript": "^4.2.4"
   }
 }
